Add files via upload
Initial version. Only contains one function overloaded to work with int, float, and double data types.

goodData usage: Pass an int, float, or double into data as a parameter, and it will ask for input from the user and check if there is input failure. If there is, then an error message is displayed and you are asked to input data again.